Put slog tests in a helper, move funcr test
