+++ funcr/slogsink_test.go
@@ -20,6 +20,7 @@ limitations under the License.
 package funcr
 
 import (
+	"bytes"
 	"fmt"
 	"log/slog"
 	"path/filepath"
@@ -27,6 +28,7 @@ import (
 	"testing"
 
 	"github.com/go-logr/logr"
+	"github.com/go-logr/logr/internal/testhelp"
 )
 
 func TestSlogSink(t *testing.T) {
@@ -106,5 +108,46 @@ func TestSlogSinkWithCaller(t *testing.T) {
 	expect := fmt.Sprintf(`{"logger":"","caller":{"file":%q,"line":%d},"msg":"msg","error":null,"int":1}`, filepath.Base(file), line-1)
 	if capt.log != expect {
 		t.Errorf("\nexpected %q\n     got %q", expect, capt.log)
+	}
+}
+
+func TestRunSlogTests(t *testing.T) {
+	fn := func(buffer *bytes.Buffer) slog.Handler {
+		printfn := func(obj string) {
+			fmt.Fprintln(buffer, obj)
+		}
+		opts := Options{
+			LogTimestamp: true,
+			Verbosity:    10,
+			RenderBuiltinsHook: func(kvList []any) []any {
+				mappedKVList := make([]any, len(kvList))
+				for i := 0; i < len(kvList); i += 2 {
+					key := kvList[i]
+					switch key {
+					case "ts":
+						mappedKVList[i] = "time"
+					default:
+						mappedKVList[i] = key
+					}
+					mappedKVList[i+1] = kvList[i+1]
+				}
+				return mappedKVList
+			},
+		}
+		logger := NewJSON(printfn, opts)
+		return logr.ToSlogHandler(logger)
+	}
+	exceptions := []string{
+		"a Handler should ignore a zero Record.Time", // Time is generated by sink.
+	}
+	testhelp.RunSlogTests(t, fn, exceptions...)
+}
+
+func TestLogrSlogConversion(t *testing.T) {
+	f := New(func(prefix, args string) {}, Options{})
+	f2 := logr.FromSlogHandler(logr.ToSlogHandler(f))
+	if want, got := f, f2; got != want {
+		t.Helper()
+		t.Errorf("Expected %T %+v, got instead: %T %+v", want, want, got, got)
 	}
 }
