+++ internal/testhelp/slog_test.go
@@ -1,0 +1,33 @@
+//go:build go1.21
+// +build go1.21
+
+/*
+Copyright 2023 The logr Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package testhelp
+
+import (
+	"bytes"
+	"log/slog"
+	"testing"
+)
+
+func TestRunSlogTestsOnSlogSink(t *testing.T) {
+	// This proves that RunSlogTests works.
+	RunSlogTests(t, func(buffer *bytes.Buffer) slog.Handler {
+		return slog.NewJSONHandler(buffer, nil)
+	})
+}
