+++ slogr_test.go
@@ -21,7 +21,6 @@ package logr_test
 
 import (
 	"bytes"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -31,10 +30,10 @@ import (
 	"runtime"
 	"strings"
 	"testing"
-	"testing/slogtest"
 
 	"github.com/go-logr/logr"
 	"github.com/go-logr/logr/funcr"
+	"github.com/go-logr/logr/internal/testhelp"
 )
 
 var debugWithoutTime = &slog.HandlerOptions{
@@ -105,92 +104,16 @@ func TestWithCallDepth(t *testing.T) {
 	}
 }
 
-func TestJSONHandler(t *testing.T) {
+func TestRunSlogTestsOnSlogSink(t *testing.T) {
 	// This proves that slogSink passes slog's own tests.
-	testSlog(t, func(buffer *bytes.Buffer) slog.Handler {
-		return slog.NewJSONHandler(buffer, nil)
-	})
-}
-
-func TestFuncrHandler(t *testing.T) {
-	fn := func(buffer *bytes.Buffer) slog.Handler {
-		printfn := func(obj string) {
-			fmt.Fprintln(buffer, obj)
-		}
-		opts := funcr.Options{
-			LogTimestamp: true,
-			Verbosity:    10,
-			RenderBuiltinsHook: func(kvList []any) []any {
-				mappedKVList := make([]any, len(kvList))
-				for i := 0; i < len(kvList); i += 2 {
-					key := kvList[i]
-					switch key {
-					case "ts":
-						mappedKVList[i] = "time"
-					default:
-						mappedKVList[i] = key
-					}
-					mappedKVList[i+1] = kvList[i+1]
-				}
-				return mappedKVList
-			},
-		}
-		logger := funcr.NewJSON(printfn, opts)
+	testhelp.RunSlogTests(t, func(buffer *bytes.Buffer) slog.Handler {
+		handler := slog.NewJSONHandler(buffer, nil)
+		logger := logr.FromSlogHandler(handler)
 		return logr.ToSlogHandler(logger)
-	}
-	exceptions := []string{
-		"a Handler should ignore a zero Record.Time", // Time is generated by sink.
-	}
-	testSlog(t, fn, exceptions...)
-}
-
-func testSlog(t *testing.T, createHandler func(buffer *bytes.Buffer) slog.Handler, exceptions ...string) {
-	var buffer bytes.Buffer
-	handler := createHandler(&buffer)
-	err := slogtest.TestHandler(handler, func() []map[string]any {
-		var ms []map[string]any
-		for _, line := range bytes.Split(buffer.Bytes(), []byte{'\n'}) {
-			if len(line) == 0 {
-				continue
-			}
-			var m map[string]any
-			if err := json.Unmarshal(line, &m); err != nil {
-				t.Errorf("%v: %q", err, string(line))
-			}
-			ms = append(ms, m)
-		}
-		return ms
 	})
-
-	// Correlating failures with individual test cases is hard with the current API.
-	// See https://github.com/golang/go/issues/61758
-	t.Logf("Output:\n%s", buffer.String())
-	if err != nil {
-		if unwrappable, ok := err.(interface {
-			Unwrap() []error
-		}); ok {
-			for _, err := range unwrappable.Unwrap() {
-				if !containsOne(err.Error(), exceptions...) {
-					t.Errorf("Unexpected error: %v", err)
-				}
-			}
-		} else {
-			// Shouldn't be reached, errors from errors.Join can be split up.
-			t.Errorf("Unexpected errors:\n%v", err)
-		}
-	}
 }
 
-func containsOne(hay string, needles ...string) bool {
-	for _, needle := range needles {
-		if strings.Contains(hay, needle) {
-			return true
-		}
-	}
-	return false
-}
-
-func TestDiscard(_ *testing.T) {
+func TestSlogSinkOnDiscard(_ *testing.T) {
 	// Compile-test
 	logger := slog.New(logr.ToSlogHandler(logr.Discard()))
 	logger.WithGroup("foo").With("x", 1).Info("hello")
@@ -204,10 +127,6 @@ func TestConversion(t *testing.T) {
 	e := logr.Logger{}
 	e2 := logr.FromSlogHandler(logr.ToSlogHandler(e))
 	expectEqual(t, e, e2)
-
-	f := funcr.New(func(prefix, args string) {}, funcr.Options{})
-	f2 := logr.FromSlogHandler(logr.ToSlogHandler(f))
-	expectEqual(t, f, f2)
 
 	text := slog.NewTextHandler(io.Discard, nil)
 	text2 := logr.ToSlogHandler(logr.FromSlogHandler(text))
